+++ libs/nest/src/lib/compliance/scoring/scoring.service.ts
@@ -203,8 +203,8 @@ export class ScoringService extends BaseWorker {
           question.max_score = 0;
         }
 
-        question.ai_answered = this.filterService.questionHasValidAnswer(response.ai_response, 'answer', question.component);
-        question.user_answered = this.filterService.questionHasValidAnswer(response.org_response, 'value', question.component);
+        question.user_answered = response.org_response ? Object.prototype.hasOwnProperty.call(response.org_response, 'value') : false;
+        question.ai_answered = !question.user_answered && this.filterService.questionHasValidAnswer(response.ai_response, 'answer', question.component);
         question.not_started = !question.ai_answered && !question.org_answered;
       }
       /**
@@ -240,7 +240,7 @@ export class ScoringService extends BaseWorker {
     }
 
     set(question, 'counts', {
-      not_started: question.not_started ? 1 : 0,
+      not_started: !question.user_answered && !question.ai_answered ? 1 : 0,
       org_answered: question.user_answered ? 1 : 0,
       ai_answered: question.ai_answered ? 1 : 0,
       bookmarked: question.bookmarked ? 1 : 0,